# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ogp.OGPData:
    properties:
      description:
        type: string
      image:
        type: string
      site_name:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
...
  usecase.PodcastDetailResult:
    properties:
      artwork_url:
        type: string
      author:
        type: string
      average_rating:
        type: number
      created_at:
        type: string
      description:
        type: string
      favorite_count:
        type: integer
      feed_url:
        type: string
      genre:
        type: string
      id:
        type: string
      title:
        type: string
      total_ratings:
        type: integer
    type: object
...
  usecase.PodcastSearchItem:
    properties:
      artwork_url:
        type: string
      author:
        type: string
      average_rating:
        type: number
      favorite_count:
        type: integer
      id:
        type: string
      title:
        type: string
      total_ratings:
        type: integer
    type: object
  usecase.PodcastSearchResult:
    properties:
      podcasts:
        items:
          $ref: '#/definitions/usecase.PodcastSearchItem'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: お笑いラジオの視聴記録SNS API
  title: Podlog API
  version: "1.0"
paths:
...
  /podcasts/{id}:
    get:
      description: ポッドキャストIDから詳細情報を取得します。平均評価・評価件数を含みます。
      parameters:
      - description: ポッドキャストID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.PodcastDetailResult'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ポッドキャスト詳細取得
      tags:
      - podcasts
...
  /podcasts/fetch-url:
    post:
      consumes:
      - application/json
      description: 外部URLからOGP情報を取得します
      parameters:
      - description: URL
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ogp.OGPData'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: URL情報取得
      tags:
      - podcasts
  /podcasts/popular:
    get:
      description: 評価件数の多い番組をランキング順で取得します。探す画面の「人気の番組」セクションで使用します。
      parameters:
      - default: 10
        description: 最大取得件数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.PodcastSearchResult'
      summary: 人気ポッドキャスト一覧
      tags:
      - podcasts
...
  /podcasts/search:
    get:
      description: アプリ内 DB に登録済みの番組をキーワードで検索します。平均評価・評価件数を含みます。genre パラメータでジャンル絞り込みが可能です。
      parameters:
      - description: 検索キーワード（genre 指定時は省略可）
        in: query
        name: q
        type: string
      - description: 親カテゴリ名（ジャンル一覧 API の id を指定。サブカテゴリに自動展開される）
        in: query
        name: genre
        type: string
      - default: 20
        description: 最大取得件数
        in: query
        name: limit
        type: integer
      - default: 0
        description: オフセット
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.PodcastSearchResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ポッドキャスト検索
      tags:
      - podcasts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"